move launch logic to its own package
