+++ connector/connector.go
@@ -4,11 +4,9 @@ import (
 	"errors"
 	"fmt"
 	"net/url"
-	"os"
-	"os/exec"
-	"strconv"
 	"strings"
 
+	"github.com/18F/cf-db-connect/launcher"
 	"github.com/18F/cf-db-connect/models"
 
 	"code.cloudfoundry.org/cli/plugin"
@@ -47,7 +45,7 @@ func Connect(cliConnection plugin.CliConnection, appName, serviceInstanceName st
 	}
 
 	fmt.Println("Setting up SSH tunnel...")
-	localPort, cmd, err := createSSHTunnel(serviceKeyCreds, appName)
+	localPort, cmd, err := launcher.CreateSSHTunnel(serviceKeyCreds, appName)
 	if err != nil {
 		return
 	}
@@ -57,13 +55,13 @@ func Connect(cliConnection plugin.CliConnection, appName, serviceInstanceName st
 
 	if isMySQLService(serviceName, planName) {
 		fmt.Println("Connecting to MySQL...")
-		err = launchMySQL(localPort, serviceKeyCreds)
+		err = launcher.LaunchMySQL(localPort, serviceKeyCreds)
 		if err != nil {
 			return
 		}
 	} else if isPSQLService(serviceName, planName) {
 		fmt.Println("Connecting to Postgres...")
-		err = launchPSQL(localPort, serviceKeyCreds)
+		err = launcher.LaunchPSQL(localPort, serviceKeyCreds)
 		if err != nil {
 			return
 		}
@@ -94,38 +92,6 @@ func getCreds(cliConnection plugin.CliConnection, serviceGUID, serviceKeyID stri
 	return
 }
 
-func createSSHTunnel(serviceKeyCreds models.Credentials, appName string) (localPort int, cmd *exec.Cmd, err error) {
-	localPort = getAvailablePort()
-	cmd = exec.Command("cf", "ssh", "-N", "-L", fmt.Sprintf("%d:%s:%s", localPort, serviceKeyCreds.Host, serviceKeyCreds.Port), appName)
-	err = cmd.Start()
-	return
-}
-
-func launchMySQL(localPort int, serviceKeyCreds models.Credentials) error {
-	fmt.Printf("%+v\n", serviceKeyCreds)
-	return startShell("mysql", []string{"-u", serviceKeyCreds.Username, "-h", "0", "-p" + serviceKeyCreds.Password, "-D", serviceKeyCreds.DBName, "-P", strconv.Itoa(localPort)})
-}
-
-func launchPSQL(localPort int, serviceKeyCreds models.Credentials) error {
-	os.Setenv("PGPASSWORD", serviceKeyCreds.Password)
-	return startShell("psql", []string{"-h", "localhost", "-p", fmt.Sprintf("%d", localPort), serviceKeyCreds.DBName, serviceKeyCreds.Username})
-}
-
-// derived from http://technosophos.com/2014/07/11/start-an-interactive-shell-from-within-go.html
-func startShell(name string, args []string) error {
-	cmd := exec.Command(name, args...)
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	// Wait until user exits the shell
-	return cmd.Run()
-}
-
-func getAvailablePort() int {
-	// TODO find one that's available
-	return 63306
-}
 func generateServiceKeyID() string {
 	// TODO find one that's available, or randomize
 	return "DB_CONNECT"
